🧚🏽‍♀️ Working more on a player selection screen
Adding ability to pick a player from the roster.

+++ src/identities.rs
@@ -4,7 +4,7 @@ use std::path::Path;
 #[derive(Debug, Default)]
 pub struct Identities {
     /// Each value is the index in the players file. Up to 31 (inclusive), as we only support 32 players.
-    pub players: [Option<usize>; 4],
+    pub players: [Option<u8>; 4],
 }
 
 impl Identities {
@@ -16,7 +16,7 @@ impl Identities {
                 let mut identities = Identities::default();
                 for (idx, player_idx) in data.iter().enumerate() {
                     if *player_idx != 0 {
-                        identities.players[idx] = Some((*player_idx as usize - 1).min(31));
+                        identities.players[idx] = Some((*player_idx - 1).min(31));
                     }
                 }
                 identities
